requirements.txt
aiofiles==24.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
colorlog==6.8.2
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Injector==0.15.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
groq==0.9.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
injector==0.21.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
pathlib==1.0.1
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
PyPDF2==3.0.1
pyperclip==1.8.2
python-docx==0.8.11
python-dotenv==1.0.1
PyYAML==6.0.1
reportlab==3.6.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
